# demoapi/api/openapi.yaml
openapi: 3.0.3
info:
  title: demoapi
  version: 1.0.0
  description: |
    A fake API to build TUIs against. Applications shaped like Argo CD's,
    jobs that run against them, AWX-shaped job events, and an
    Elasticsearch-shaped log index. See the demoapi package doc.

    Paging, filtering and sorting are plain parameters. Nothing here
    declares how a client should page: that belongs to the client (for a
    tuilib app, its design file), not to the spec.

    Every operation honours the failure knobs ?latency=, ?fail= and
    ?flaky=, applied before routing.

    It is a fixture. It has no authentication. Do not deploy it.
servers:
  - url: http://127.0.0.1:8099
paths:
...
  /jobs/{id}:
    parameters:
      - $ref: "#/components/parameters/jobID"
      - $ref: "#/components/parameters/latency"
      - $ref: "#/components/parameters/fail"
      - $ref: "#/components/parameters/flaky"
    get:
      operationId: getJob
      summary: One job, by the id a launch returned.
      responses:
        "200":
          description: The job.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/Job"}
        "404": {$ref: "#/components/responses/Error"}
        default: {$ref: "#/components/responses/Error"}
  /jobs/{id}/log:
    parameters:
      - $ref: "#/components/parameters/jobID"
      - $ref: "#/components/parameters/latency"
      - $ref: "#/components/parameters/fail"
      - $ref: "#/components/parameters/flaky"
    get:
      operationId: getJobLog
      summary: The job's output, streamed a line at a time, then closed.
      parameters:
        - {name: rate, in: query, schema: {type: string}, description: "Delay between lines. Duration or bare milliseconds; default 150ms."}
      responses:
        "200":
          description: Chunked plain text, flushed per line.
          content:
            text/plain:
              schema: {type: string}
        "404": {$ref: "#/components/responses/Error"}
        default: {$ref: "#/components/responses/Error"}
...
  /api/v2/jobs/{id}/:
    parameters:
      - $ref: "#/components/parameters/awxJobID"
      - $ref: "#/components/parameters/latency"
      - $ref: "#/components/parameters/fail"
      - $ref: "#/components/parameters/flaky"
    get:
      operationId: awxGetJob
      summary: One AWX-shaped job.
      responses:
        "200":
          description: The job.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/AwxJob"}
        "404": {$ref: "#/components/responses/Error"}
        default: {$ref: "#/components/responses/Error"}
...
components:
  parameters:
    latency:
      name: latency
      in: query
      description: Hold the request this long before answering. Duration or bare milliseconds.
      schema: {type: string}
    fail:
      name: fail
      in: query
      description: Answer with this status code (>= 400) instead.
      schema: {type: integer}
    flaky:
      name: flaky
      in: query
      description: Probability (0–1) of answering 502.
      schema: {type: number}
...
    jobID:
      name: id
      in: path
      required: true
      schema: {type: string}
    awxJobID:
      name: id
      in: path
      required: true
      schema: {type: integer}
  responses:
    Error:
      description: Any failure.
      content:
        application/json:
          schema: {$ref: "#/components/schemas/Error"}
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error: {type: string}
...
    JobKind:
      type: string
      enum: [sync, fail]
    JobStatus:
      type: string
      enum: [running, succeeded, failed]
    Job:
      type: object
      required: [id, app, kind, status, started]
      properties:
        id: {type: string}
        app: {type: string}
        kind: {$ref: "#/components/schemas/JobKind"}
        status: {$ref: "#/components/schemas/JobStatus"}
        started: {type: string, format: date-time}
        finished: {type: string, format: date-time, description: "Absent while running."}
...
    AwxJobStatus:
      type: string
      enum: [running, successful]
    AwxJob:
      type: object
      required: [id, name, status, started, event_processing_finished]
      properties:
        id: {type: integer}
        name: {type: string}
        status: {$ref: "#/components/schemas/AwxJobStatus"}
        started: {type: string, format: date-time}
        finished: {type: string, format: date-time, description: "Absent while running."}
        event_processing_finished:
          type: boolean
          description: Trails the job's status; events can still arrive after it finishes.
...